Removed all osgTerrain classes, and merged changes:
From Colin McDonald, "I was getting crashes in DataSet::SourceData::readImage if one of
destHeight or destWidth came out to be one.  There was also an error
with alpha blending in that routine.

Between osg versions 1.0 and 1.2 last year there was a large increase in
the size of many database files produced by osgdem, of the order of x2
to x4.  Some other people have commented on it in osg-users, as with
databases which were already large it has caused some problems in
transferring them to customers.

I have done some investigation into the Simplifier changes during that
period, and the principal cause appears to be a change in the logic of
osgUtil::Simplifier::continueSimplificationImplementation. Previously in
osg v1.0 it always carried on simplifying while the simplification error
was less than getMaximumError().  But in osg v1.2 it stops simplifying
whenever the max error is exceeded OR the target primitive ratio is
reached. A simple adjustment to the targetMaxNumVertices in
DataSet::DestinationTile::createPolygonal reduces the database sizes
near to the osg v1.0 sizes."



git-svn-id: http://www.openscenegraph.org/svn/VirtualPlanetBuilder/trunk@620 1bf02a86-792a-0410-856e-ea5eef037d3d