From Christophe Loustaunau, "I have test vpbmaster on windows 2000, and I think that I have have an error :
If vpb have to create a directory like : "c:\data\mars" it fails. In fact when vpb have to do that, it check all the directory possible like this :
"c:"
"c:\data"
"c:\data\mars"

But when vpb want's to know if "c:" exist with osgDB::FileType type = osgDB::fileType(path) It fails,
It's because "c:" doesn't exist but "c:\" does.


So here is a patch based on the svn, which correct that."


git-svn-id: http://www.openscenegraph.org/svn/VirtualPlanetBuilder/trunk@912 1bf02a86-792a-0410-856e-ea5eef037d3d