From Christophe Loustaunau,
First submission: "But with vpbmaster I have a little problem :

If I use the option --run-path d:\tmp
vpbmaster write in the task files : --run-path C:\Users\Christophe 

It seems to use the default path and not the path I have specified.
I have just added the line

taskManager->setRunPath(runPath);

when vpbMaster read the --run-path option and it works. I don't know if it's the better solution but anyway
here is the file modified vpbmaster.cpp"


Second submission: "If osg is compiled with OSG_USE_FLOAT_BOUNDINGSPHERE = OFF
There is compile errors with VPB :

osg::maximum(cutOffDistance, pagedLOD->getBound().radius()*_dataSet->getRadiusToMaxVisibleDistanceRatio());

cutOffDistance is a float and the other part a double.

Here is the file Destination.cpp modified"


git-svn-id: http://www.openscenegraph.org/svn/VirtualPlanetBuilder/trunk@946 1bf02a86-792a-0410-856e-ea5eef037d3d